- this has changed in the daemon, update here too.

+++ msn/src/msnpacket.h
@@ -36,7 +36,7 @@ public:
   const unsigned short Command() { return 0; }
   const unsigned short SubSequence() { return 0; }
   const unsigned short SubCommand() { return 0; }
-  const unsigned long Sequence() { return m_nSequence; }
+  const unsigned short Sequence() { return m_nSequence; }
   const char* MSNCommand() { return m_szCommand; }
   
   virtual void InitBuffer();
@@ -45,7 +45,7 @@ protected:
   
   CMSNBuffer *m_pBuffer;
   char *m_szCommand;
-  unsigned long m_nSequence;
+  unsigned short m_nSequence;
   unsigned long m_nSize;
   bool m_bPing;
